extend interpreter base class with platform methods
