+++ packages/interpreter/.gitignore
@@ -1,0 +1,1 @@
+gen2/
