+++ packages/interpreter/src/ts/form.ts
@@ -1,0 +1,3 @@
+function retrieveFormValues() {
+
+}
