+++ packages/interpreter/tests/e2e.rs
@@ -1,0 +1,1 @@
+//! Ensure that the interpreter loads, has no errors, and writes mutations
