+++ packages/interpreter/tests/serialize.rs
@@ -1,0 +1,1 @@
+//! Ensure that events are serialized and deserialized correctly.
