+++ packages/interpreter/tsconfig.json
@@ -1,0 +1,20 @@
+{
+    "compilerOptions": {
+        "module": "system",
+        "lib": [
+            "ES2015",
+            "DOM"
+        ],
+        "noImplicitAny": true,
+        "removeComments": true,
+        "preserveConstEnums": true,
+        "outFile": "gen/main.js",
+        "sourceMap": true
+    },
+    "include": [
+        "src/ts/*.ts"
+    ],
+    "exclude": [
+        "**/*.spec.ts"
+    ]
+}
